fetch_data/requirements.txt
beautifulsoup4==4.12.3
pdf2image==1.17.0
pytesseract==0.3.13
requests==2.32.3
pillow==10.4.0
opencv-python==4.10.0.84
numpy==1.26.4
aiohttp==3.10.7
aiofiles==24.1.0